- storing usedcapacity at commitblobberconnection -

+++ code/go/0chain.net/smartcontract/partitions/blobber_challenge.go
@@ -15,8 +15,8 @@ import (
 //------------------------------------------------------------------------------
 
 type BlobberChallengeNode struct {
-	BlobberID    string  `json:"blobber_id"`
-	UsedCapacity float64 `json:"used_capacity"`
+	BlobberID    string `json:"blobber_id"`
+	UsedCapacity int64  `json:"used_capacity"`
 }
 
 func (bcn *BlobberChallengeNode) Encode() []byte {
